requirements.txt
agate==1.7.1
async-timeout==3.0.1
attrs==23.2.0
Babel==2.14.0
botocore==1.34.73
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.5
cryptography==42.0.5
cycler==0.12.1
dbt-bigquery==1.7.3
dbt-core==1.7.5
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.4
distlib==0.3.8
entrypoints==0.4
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
future==1.0.0
gitdb==4.0.11
google-api-core==2.15.0
google-auth==2.29.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.9.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
grpcio-tools==1.48.2
hologram==0.0.16
humanize==4.9.0
identify==2.5.35
idna==3.6
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
leather==0.4.0
Logbook==1.5.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mashumaro==3.12
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==8.14.0
msgpack==1.0.8
multidict==6.0.5
networkx==2.8.8
nodeenv==1.8.0
numpy==1.26.4
nvsmi==0.4.2
packaging==21.3
pandas==1.3.5
parsedatetime==2.4
pathspec==0.11.2
pillow==10.2.0
platformdirs==4.2.0
polling==0.3.2
pre-commit==3.3.3
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==7.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==1.10.14
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pyrsistent==0.20.0
python-dateutil==2.8.2
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
text-unidecode==1.3
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==1.26.18
virtualenv==20.25.1
Werkzeug==2.1.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
